feat: Number and Random are no longer rely on Config

+++ test/Timer.h
@@ -1,17 +1,18 @@
 #ifndef MTK_TEST_TIMER_H
 #define MTK_TEST_TIMER_H
 
-#include "../src/Config.h"
+#include <cstddef>
+#include <chrono>
 
 namespace mtk
 {
-    Int timer();
+    size_t timer();
 
-    inline Int timer()
+    inline size_t timer()
     {
         static std::chrono::_V2::steady_clock::time_point time = std::chrono::steady_clock::now();
         std::chrono::_V2::steady_clock::time_point now = std::chrono::steady_clock::now();
-        Int duration = std::chrono::duration_cast<std::chrono::milliseconds>(now - time).count();
+        size_t duration = std::chrono::duration_cast<std::chrono::milliseconds>(now - time).count();
         time = now;
         return duration;
     }
